requirements.txt
Flask==3.0.3
Werkzeug>=3.0.0
numpy
opencv-python
Pillow
torch
torchvision
timm
albumentations
scikit-learn
matplotlib
tqdm
python-dotenv
requests
transformers>=4.30